+++ appcheck/firebase-appcheck-playintegrity/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 19.1.0
+
 - [fixed] Fixed issue preventing limited use tokens from being correctly generated. (#8204)
 
 # 19.0.2