+++ firebase-dataconnect/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 17.3.0
+
 - [feature] `QuerySubscription.flow` gains
   [realtime updates](https://firebase.google.com/docs/sql-connect/realtime).
   ([#8186](https://github.com/firebase/firebase-android-sdk/pull/8186))
@@ -12,6 +14,9 @@
 - [fixed] Queries executed with FetchPolicy.CACHE_ONLY now fail, as expected,
   if local caching is not enabled, instead of behaving like SERVER_ONLY.
   ([#8214](https://github.com/firebase/firebase-android-sdk/pull/8214))
+- [changed] Realtime query results now update the local cache as query
+  results are received.
+  ([#8220](https://github.com/firebase/firebase-android-sdk/pull/8220))
 
 # 17.2.2